resolver/requirements.txt
pyyaml
requests
python-dotenv
fastapi
uvicorn
web3
anthropic
